Implemented Feature #2
at each __init__ of a Mail object, KeyBiz checks the availability of
the uid at the key server and may download the key if a signature
might be missing. If a signed uid is found, the key will be imported.

The GPGCommand method is now static, because it is not specific to any
GPG key at all.